source/MLLM/smollm-main/tools/smol_tools/requirements.txt
tkmacosx>=1.0.5
pynput>=1.7.7
llama-cpp-python>=0.3.1
pyperclip>=1.9.0
transformers>=4.46.2
pygments>=2.18.0